Remove sysfs gpio dependency and use gpiochips
Deprecated several tools that are not needed with new kernels.

gpiolib/gpioctl are removed and gpiod should be used instead.
adc8390 is replaced with iio in the later TS-4900+TS-TPC-8390 kernels
nvramctl is replaced with an eeprom driver in newer kernels, eg:
  /sys/devices/platform/soc/2100000.bus/21a0000.i2c/i2c-0/0-0057/eeprom
load_fpga is not used by third parties, but has a replacement standalone tool if needed.