tsmicroctl: fix sleep prep on TS-7990
Previously, the check for the pixcir controller would always
succeed and would generate a lot of noise when the commands to
write the controller would fail.

Modified to use existing write functions and do a correct poll for
the pixcir controller. This will still generate a single line of
noise on the resistive touchscreen, but will not error out.

Signed-off-by: Kris Bahnsen <kris@embeddedTS.com>